% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_season.R
\name{append_season_model}
\alias{append_season_model}
\title{Append Season Model}
\usage{
append_season_model(model = model, var = var, data = data)
}
\arguments{
\item{model}{model object}

\item{var}{string}

\item{data}{data frame}
}
\value{

}
\description{
Helper for find_season()
}
