% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_df.R
\name{bank_df}
\alias{bank_df}
\title{AdBank Data Frame}
\usage{
bank_df(DF)
}
\arguments{
\item{DF}{data.frame to be adbanked}
}
\value{

}
\description{
Creates all AdBanks in \code{c(seq(1:9), seq(from= 10, to = 90, by = 10))} for
a given data frame.
}
\details{
ATTENTION: function index is unstable (excel vs. csv file source). Always
check for correct behavior after usage!
--> Malfunction might be attributed to bank_it function
}
