% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank_it.R
\name{bank_it}
\alias{bank_it}
\title{Bank It}
\usage{
bank_it(vect, adbank)
}
\arguments{
\item{vect}{vector/list? of values to be adbanked}

\item{adbank}{adbank value}
}
\value{
vector of adbanked values
}
\description{
Helper for \code{bank_df()}.
}
\details{
ATTENTION: function index is unstable (excel vs. csv file source). Always
check for correct behavior after usage!
}
