% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_season.R
\name{compare_season}
\alias{compare_season}
\title{Compare Season}
\usage{
compare_season(model, data, teiler)
}
\arguments{
\item{model}{model object}

\item{data}{data frame}

\item{teiler}{double, 8.304 for weekly}
}
\value{

}
\description{
Gives an overview of p-values for all possible sin-curves
}
