% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummynate.R
\name{dummynate}
\alias{dummynate}
\title{Dummynate}
\usage{
dummynate(model)
}
\arguments{
\item{model}{model object}
}
\value{
vector of integers
}
\description{
Dummynate a given model (like dominate, get it?... get it?...) by setting a
multi-step Dummy based on std. residuals. Imitates the behavior of the dummy
function of the modeling suite by Prof. Dr. Alexander Preuss.
}
