% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_season.R
\name{find_season}
\alias{find_season}
\title{Find Season}
\usage{
find_season(model, data, teiler)
}
\arguments{
\item{model}{model object}

\item{data}{data frame}

\item{teiler}{double, 8.304 for weekly}
}
\value{

}
\description{
Gives an overview of p-values for all possible sin-curves
}
