% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_sig.R
\name{get_season_sig}
\alias{get_season_sig}
\title{Extract p-Value for Season Variable}
\usage{
get_season_sig(model, var)
}
\arguments{
\item{model}{model object}

\item{var}{string}
}
\value{

}
\description{
Helper for find_season()
}
