% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig.R
\name{get_sig}
\alias{get_sig}
\title{Get p-value}
\usage{
get_sig(model, var)
}
\arguments{
\item{model}{model object}

\item{var}{string, variable name}
}
\value{

}
\description{
Extracts the p-value of a variable from a model
}
