% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_adhoc.R
\name{load_adhoc}
\alias{load_adhoc}
\title{Load Ad-hoc MMM Files}
\usage{
load_adhoc(MatrixA, Modellgleichung)
}
\arguments{
\item{MatrixA}{.MXA data file from ad-hoc modeling-script}

\item{Modellgleichung}{.mgl model file from ad-hoc modeling-script}
}
\value{
prints out a summary of the model and exports data as "daten"and the variables: time_series, av, zielvar, modellvar, warnings_notFound, warnings_noVariance
}
\description{
Load Ad-hoc MMM Files
}
