% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_suffix.R
\name{log_suffix}
\alias{log_suffix}
\title{Log Suffix}
\usage{
log_suffix(x)
}
\arguments{
\item{x}{data frame to log}
}
\value{
data frame of logged values
}
\description{
Modified \code{logger()} function (by B. Kreutzer, see package maRketingscience)
to add a "_log" suffix instead of a prefix to better match other functions in
montelo and byndmodfun packages.
}
