% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_geocode.R
\name{map_geocode}
\alias{map_geocode}
\title{Get GeoCode Information from Google-Maps}
\usage{
map_geocode(data)
}
\arguments{
\item{data}{data.frame with a column 'addresses' which should contain the street, house number and city separated by spaces}
}
\value{
returns a data.frame with the additional columns 'lon', 'lat' and 'geoAddresses'
}
\description{
This function requires an API key from Google Cloud to work. The key needs to be registered on first use inside R (or is written directly into .Renviron)
}
\details{
Note that in case of an error, the geoAddress column will repeat it's first valid entry (this should be fixed in a future update)
}
