% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_plot.R
\name{mmm_plot}
\alias{mmm_plot}
\title{MMM Plot}
\usage{
mmm_plot(model, av, time_series)
}
\arguments{
\item{model}{model object}

\item{av}{AV of the model}

\item{time_series}{time series object for x-axis}
}
\value{

}
\description{
Produces a time series plot to compare model prediction and actual values for
a given AV. Also displays residuals in an interactive display in the viewer.
}
