% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_template.R
\name{mmm_template}
\alias{mmm_template}
\title{MMM Template}
\usage{
mmm_template(path = NULL, filename = "mmm_template")
}
\arguments{
\item{path}{string, optional path to save the file to, default is current wd}

\item{filename}{string for the file name e.g. "myModelFile" - no file extension}
}
\value{
Generates a .R-file with a template for MMM-Modeling
}
\description{
Generates a .R-file with the basic code structure and dependencies for MMM.
}
\details{
Default library for model functions is montelo
}
