% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_check.R
\name{p_check}
\alias{p_check}
\title{Look up p values for variable groups}
\usage{
p_check(varGroup, vergl = FALSE, model, data)
}
\arguments{
\item{varGroup}{string to identify variable group to be tested into model}

\item{vergl}{logical, should results be compared with other model variables?}

\item{model}{model object of interest}

\item{data}{data containing model variables and variables of varGroup}
}
\value{
plot of resulting p values
}
\description{
Looks up the p values for every member of a variable group if they were
included in a given model and plots the results. Optionally with comparison
of p values from all other variables in the given model for every instance.
}
