% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_profile_template.R
\name{r_profile_template}
\alias{r_profile_template}
\title{Create a .Rprofile file}
\usage{
r_profile_template(path = NULL, filename = "removethis")
}
\arguments{
\item{path}{file path in which to create the file (defaults to current directory)}

\item{filename}{placeholder to prevent the .Rprofile from becoming active right away. Needs to be removed (filename = '.Rprofile')}
}
\description{
Creates a .Rprofile file in the chosen directory
}
