% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_adhoc.R
\name{save_adhoc}
\alias{save_adhoc}
\title{Save Ad-hoc}
\usage{
save_adhoc(daten, model, av_name, sin_schleife = 51, sin_teiler = 8.3042502)
}
\arguments{
\item{daten}{data.frame of the data}

\item{model}{model object}

\item{av_name}{string, name of the dependet variable}

\item{sin_schleife}{length of sin loop, default = 51}

\item{sin_teiler}{constant, default = 8.3042502 for weekly data}
}
\value{
none, saves two files to current wd
}
\description{
Save model files and data in a format needed for use in ad-hoc modeling script
}
\details{
Maybe in the future, rearrange the dataset, so that the second column contains the av, because the script expects this to be the av. As of now, be aware to change the av if the adhoc script is used
}
