% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonate.R
\name{seasonate}
\alias{seasonate}
\title{Seasonate}
\usage{
seasonate(x, av, index)
}
\arguments{
\item{x}{data frame}

\item{av}{AV}

\item{index}{Index in \code{sin((ZAEHLER[k]+index)/NENNER)}}
}
\value{
vector
}
\description{
Creates a vector which can mimic the season function of of the modeling
suite by Prof. Dr. Alexander Preuss for comparability and reproduction of
models
}
