% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonate_optimize.R
\name{seasonate_optimize}
\alias{seasonate_optimize}
\title{Find optimal Season}
\usage{
seasonate_optimize(x, av)
}
\arguments{
\item{x}{data frame}

\item{av}{vector defining the AV}
}
\value{
list with a vector for Sinus_Saison and the value of the constant used
}
\description{
Finds the constant for the best fitting sin-curve for a given AV. Mimics the
season function of of the modeling suite by Prof. Dr. Alexander Preuss for
comparability and reproduction of models. --- assumes weekly data
}
