% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummynate.R
\name{dummynate}
\alias{dummynate}
\title{Dummynate}
\usage{
dummynate(model)
}
\arguments{
\item{model}{model object}
}
\value{
vector of integers
}
\description{
For demonstration only. DO NOT USE!
}
\details{
Dummynate a given model (like dominate, get it?... get it?...) by setting a
multi-step Dummy based on std. residuals.
}
