% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm_plot.R
\name{mmm_plot}
\alias{mmm_plot}
\title{MMM Plot}
\usage{
mmm_plot(model, data = NULL, index = 1)
}
\arguments{
\item{model}{object of class "lm" containing the results returned by the lm() model fitting function}

\item{data}{an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{index}{single value of type "integer" or "character" to identify the time series in the original data for the model}
}
\description{
Generates a dygraphs-plot to compare model prediction and actual values for
a given AV. Also displays residuals in an interactive display in the viewer.
}
